+++ includes/Autoload.php
@@ -201,7 +201,7 @@ class Autoload {
 		require_once 'Ajax.php';
 		require_once 'class-mahalle-local.php';
 		require_once 'Hezarfen_Install.php';
-		require_once 'hezarfen-wc-helpers.php';
+		require_once 'class-hezarfen-wc-helper.php';
 
 		if ( is_admin() ) {
 			require_once 'admin/order/OrderDetails.php';
